Do not allow ClearApplicationDataForProvider to run in prod

+++ app/services/vendor_api/clear_application_data_for_provider.rb
@@ -1,6 +1,8 @@
 module VendorAPI
   class ClearApplicationDataForProvider
     def self.call(provider)
+      raise 'This is not meant to be run in production' if HostingEnvironment.production?
+
       scope = Candidate.joins(application_forms: { application_choices: { course_option: :course } })
       scope.where("courses.accredited_provider": provider).or(scope.where("courses.provider": provider))
         .delete_all
